+++ src/pages/about-page/AboutPage.tsx
@@ -1,19 +1,20 @@
 import React from "react";
 import OverlayPage from "../components/overlay/OverlayPage";
+import './AboutPage.css';
 
 const AboutPage = () => {
   return (
     <OverlayPage>
-      <div>
-        <h1 style={{ fontSize: '3rem', marginBottom: '2rem', color: '#4da6ff' }}>
+      <div className="about-page">
+        <h1>
           About Me
         </h1>
-        <div style={{ fontSize: '1.2rem', lineHeight: '1.8', maxWidth: '800px' }}>
-          <p style={{ marginBottom: '1.5rem' }}>
+        <div className="content">
+          <p>
             Welcome! I&apos;m a passionate software developer with expertise in modern web technologies 
             and a love for creating innovative digital experiences.
           </p>
-          <p style={{ marginBottom: '1.5rem' }}>
+          <p>
             My journey in technology spans across full-stack development, with particular strengths 
             in React, TypeScript, and Three.js for creating immersive web experiences.
           </p>
